requirements.txt
setuptools
gcsfs
google-cloud-storage
pybaseball
pandas
numpy
